# src/main/java/com/cy/fashion/product/dto/request/ProductCreateRequest.java
package com.cy.fashion.product.dto.request;

import lombok.AllArgsConstructor;

@AllArgsConstructor
public class ProductCreateRequest {
    public String brandId;
    public String category;
    public int price;
}